Move Sidebar into controls project
