+++ src/Files.App.Controls/Sidebar/SidebarItem.Properties.cs
@@ -1,10 +1,7 @@
 // Copyright (c) Files Community
 // Licensed under the MIT License.
 
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Controls;
-
-namespace Files.App.UserControls.Sidebar
+namespace Files.App.Controls
 {
 	public sealed partial class SidebarItem : Control
 	{
