+++ src/Files.App.Controls/Sidebar/SidebarView.xaml
@@ -1,10 +1,9 @@
 ﻿<!--  Copyright (c) Files Community. Licensed under the MIT License.  -->
 <UserControl
-	x:Class="Files.App.UserControls.Sidebar.SidebarView"
+	x:Class="Files.App.Controls.SidebarView"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-	xmlns:helpers="using:Files.App.Helpers"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	HorizontalAlignment="Stretch"
 	Loaded="SidebarView_Loaded"
@@ -118,7 +117,6 @@
 			Visibility="{Binding ElementName=PaneRoot, Path=Visibility, Mode=OneWay}">
 			<UserControl
 				x:Name="SidebarResizerControl"
-				AutomationProperties.Name="{helpers:ResourceString Name=ResizeElementControl/AutomationProperties/Name}"
 				IsTabStop="True"
 				KeyDown="SidebarResizerControl_KeyDown"
 				UseSystemFocusVisuals="True" />
