+++ src/Files.App/Data/Items/FileTagItem.cs
@@ -53,5 +53,7 @@ namespace Files.App.Data.Items
 		public object ToolTip => Text;
 
 		public bool IsExpanded { get => false; set { } }
+
+		public bool PaddedItem => false;
 	}
 }
