+++ src/Files.App/Data/Items/WslDistroItem.cs
@@ -61,6 +61,8 @@ namespace Files.App.Data.Items
 			};
 		}
 
+		public bool PaddedItem => false;
+
 		public int CompareTo(INavigationControlItem other) => Text.CompareTo(other.Text);
 	}
 }
