+++ src/Files.App/Helpers/Win32/Win32Helper.WindowManagement.cs
@@ -43,24 +43,6 @@ namespace Files.App.Helpers
 		}
 
 		/// <summary>
-		/// Sets cursor when hovering on a specific element.
-		/// </summary>
-		/// <param name="uiElement">An element to be changed.</param>
-		/// <param name="cursor">Cursor to change.</param>
-		public static void ChangeCursor(this UIElement uiElement, InputCursor cursor)
-		{
-			Type type = typeof(UIElement);
-
-			type.InvokeMember(
-				"ProtectedCursor",
-				BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.SetProperty | BindingFlags.Instance,
-				null,
-				uiElement,
-				[cursor]
-			);
-		}
-
-		/// <summary>
 		/// Force window to stay at bottom of other upper windows.
 		/// </summary>
 		/// <param name="lParam">The lParam of the message.</param>
