+++ src/Files.App/ViewModels/UserControls/SidebarViewModel.cs
@@ -1,8 +1,8 @@
 // Copyright (c) Files Community
 // Licensed under the MIT License.
 
+using Files.App.Controls;
 using Files.App.Helpers.ContextFlyouts;
-using Files.App.UserControls.Sidebar;
 using Microsoft.UI.Input;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
